Corrected the .cpp and .h file names
