+++ data-a01/.vs/data-a01/v17/DocumentLayout.backup.json
@@ -43,8 +43,8 @@
               "Title": "main.cpp",
               "DocumentMoniker": "C:\\Source\\DataStructures\\data-a01\\main.cpp",
               "RelativeDocumentMoniker": "main.cpp",
-              "ToolTip": "C:\\Source\\DataStructures\\data-a01\\main.cpp*",
-              "RelativeToolTip": "main.cpp*",
+              "ToolTip": "C:\\Source\\DataStructures\\data-a01\\main.cpp",
+              "RelativeToolTip": "main.cpp",
               "ViewState": "AgIAAF0AAAAAAAAAAAAAAHIAAAAAAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000677|",
               "WhenOpened": "2025-02-04T21:41:22.953Z",
