+++ data-a01/.vs/data-a01/v17/DocumentLayout.json
@@ -3,12 +3,12 @@
   "WorkspaceRootPath": "C:\\Source\\DataStructures\\data-a01\\",
   "Documents": [
     {
-      "AbsoluteMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|C:\\Source\\DataStructures\\data-a01\\main.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
-      "RelativeMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|solutionrelative:main.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
+      "AbsoluteMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|C:\\Source\\DataStructures\\data-a01\\LibraryCatalog.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
+      "RelativeMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|solutionrelative:LibraryCatalog.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
     },
     {
-      "AbsoluteMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|C:\\Source\\DataStructures\\data-a01\\management.h||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
-      "RelativeMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|solutionrelative:management.h||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
+      "AbsoluteMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|C:\\Source\\DataStructures\\data-a01\\LibraryCatalog.h||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
+      "RelativeMoniker": "D:0:0:{5393A8FF-75D0-4508-AE88-D84D30E88C58}|data-a01.vcxproj|solutionrelative:LibraryCatalog.h||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
     }
   ],
   "DocumentGroupContainers": [
@@ -27,25 +27,25 @@
             {
               "$type": "Document",
               "DocumentIndex": 1,
-              "Title": "management.h",
-              "DocumentMoniker": "C:\\Source\\DataStructures\\data-a01\\management.h",
-              "RelativeDocumentMoniker": "management.h",
-              "ToolTip": "C:\\Source\\DataStructures\\data-a01\\management.h",
-              "RelativeToolTip": "management.h",
-              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAA==",
+              "Title": "LibraryCatalog.h",
+              "DocumentMoniker": "C:\\Source\\DataStructures\\data-a01\\LibraryCatalog.h",
+              "RelativeDocumentMoniker": "LibraryCatalog.h",
+              "ToolTip": "C:\\Source\\DataStructures\\data-a01\\LibraryCatalog.h",
+              "RelativeToolTip": "LibraryCatalog.h",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAgAAAALAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000680|",
-              "WhenOpened": "2025-02-04T22:09:01.445Z",
+              "WhenOpened": "2025-02-05T03:42:01.489Z",
               "EditorCaption": ""
             },
             {
               "$type": "Document",
               "DocumentIndex": 0,
-              "Title": "main.cpp",
-              "DocumentMoniker": "C:\\Source\\DataStructures\\data-a01\\main.cpp",
-              "RelativeDocumentMoniker": "main.cpp",
-              "ToolTip": "C:\\Source\\DataStructures\\data-a01\\main.cpp",
-              "RelativeToolTip": "main.cpp",
-              "ViewState": "AgIAAF0AAAAAAAAAAAAAAHIAAAAAAAAAAAAAAA==",
+              "Title": "LibraryCatalog.cpp",
+              "DocumentMoniker": "C:\\Source\\DataStructures\\data-a01\\LibraryCatalog.cpp",
+              "RelativeDocumentMoniker": "LibraryCatalog.cpp",
+              "ToolTip": "C:\\Source\\DataStructures\\data-a01\\LibraryCatalog.cpp",
+              "RelativeToolTip": "LibraryCatalog.cpp",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAoAAAAAAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000677|",
               "WhenOpened": "2025-02-04T21:41:22.953Z",
               "EditorCaption": ""
