+++ data-a01/data-a01.vcxproj
@@ -127,10 +127,10 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp" />
+    <ClCompile Include="LibraryCatalog.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="management.h" />
+    <ClInclude Include="LibraryCatalog.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
