+++ data-a01/data-a01.vcxproj.filters
@@ -15,12 +15,12 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp">
+    <ClCompile Include="LibraryCatalog.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="management.h">
+    <ClInclude Include="LibraryCatalog.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
